Move epinio helper package to acceptance
This move is made since this package is used in the helper directory
